Updating board after move. Position evaluation. MinMax algorithm in a basic form.

+++ engine/Project1/PositionEvaluator.h
@@ -6,5 +6,5 @@ class PositionEvaluator
 {
 public:
 
-	virtual double Evaluate(const Position& game) = 0;
+	virtual int32_t Evaluate(const Position& game) = 0;
 };
