+++ engine/Project1/StandardPositionEvaluator.h
@@ -6,6 +6,6 @@ class StandardPositionEvaluator : public PositionEvaluator
 {
 public:
 
-	static std::unordered_map <Piece::Type, double> piecesMaterial;
-	virtual double Evaluate(const Position& game);
+	static std::unordered_map <Piece::Type, uint32_t> piecesMaterial;
+	virtual int32_t Evaluate(const Position& game);
 };
